auth/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.16
asyncpg==0.27.0
python-jose==3.3.0
psycopg2-binary==2.9.6
pika==1.3.1
passlib[bcrypt]==1.7.4 